refactor(spp): split nullifier_batch_update.go into focused files
Mirror the transact-circuit layout. nullifier_batch_update.go keeps
only the circuit (type, constructor, Define) and setup/compile. The rest
moves to focused files: bundle IO + DTOs to
nullifier_batch_update_bundle.go, assignment building to
nullifier_batch_update_assignment.go, and the indexed-tree batch-insert
witness to indexed_tree_batch.go. The queue hash-chain joins the other
hash primitives: QueueHashChain in hash.go, QueueHashChainCircuit in
hash_circuit.go. No behavior change.
